+++ src/app/main/page.tsx
@@ -1,9 +1,11 @@
 import { Content } from './Content';
+import type { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: '대시보드',
+  description: '각종 Velog 통계를 볼 수 있는 대시보드',
+};
 
 export default function Home() {
-  return (
-    <>
-      <Content />
-    </>
-  );
+  return <Content />;
 }
